admin/requirements/dev.txt
-r base.txt
isort==5.6.4
django-extensions==3.0.9
bpython==0.20.1